requirements.txt
# web_rag_engine/requirements.txt

# Web Framework
fastapi
uvicorn[standard]

# Task Queue
celery
redis

# Database
sqlalchemy
psycopg2-binary       # PostgreSQL driver
alembic               # For DB migrations (good practice)

# Vector DB Client
qdrant-client

# Web Scraping & Content
beautifulsoup4
requests
newspaper3k
lxml[html_clean] 

# ML & AI
python-dotenv         # For loading .env files
pydantic-settings     # For type-safe settings management
groq                      # Groq API client
sentence-transformers    # For embedding generation 

